refactor(rsc): clarify cache dependency filenames
