+++ packages/plugin-rsc/examples/use-cache-persistent/src/features/file-directive-from-server/dep-direct.ts
@@ -1,4 +1,4 @@
-import { transitiveLabel } from './transitive'
+import { transitiveLabel } from './dep-transitive'
 
 export function getDirectLabel() {
   return `direct-v1 + ${transitiveLabel}`
