+++ packages/plugin-rsc/examples/use-cache-persistent/src/features/file-directive-from-server/dep-transitive.ts
